Move EventFeature constants into separate interface
This patch moves the EventFeature event constants into a separate interface,
removing the need to add a hard dependency on zend-eventmanager to the
component. `EventFeature` now implements the interface, ensuring that references
to the constants will continue to work, while `AbstractTableGateway` now uses
the interface constants instead (ensuring that the zend-eventmanager dependency
is not raised during usage).