Implement FeatureSet canCallMagicCall and callMagicCall methods
Recently, I was attempting to call the lastSequenceId method of the
SequenceFeature, after adding the SequenceFeature to a TableGateway
object. However, on doing so, I discovered that both methods effectively
weren't implemented, despite what the manual alluded to. Given that I'm
adding this commit, along with the covering tests to implement the
functionality.