Refactor Insert::values() method to properly merge values with columns
When attempting to use both columns() and values() together to set the
columns and value for an insert object, and when passing non-associative
arrays to both, as seems logically correct to be able to do,
the generated SQL query has the numeric keys of the values array
overwrite the columns set from the array passed to columns().
What I would expect is that the two were properly combined. This minor
fix does that, and leaves the remaining functionality in place.