Allow tests to run when no databases are setup
Prior to 2.7.1, tests ran when:

- ext/pgsql was enabled
- but no databases were setup for pgsql

With the changes introduced to resolve #70, however, the above
combination causes cascading failures across the test suite.

Marking the two tests introduced to run in separate processes largely
solves the issues. Updating the first test to test for a connection
exception prevents a test failure when no database is setup.