Update to use zend-hydrator v3 classes when available
One major change in zend-hydrator v3 is that we have added "Hydrator"
suffixes to all hydrators. The original versions still exist, and can be
used, but emit deprecation notices.

What this change does is version detection: it checks for the new class
name, and uses that as the default if available, falling back to the old
class names. This prevents the deprecation notices, while proving that
both versions work fine with this component.