MysqlMetadata Enhancements
Add support for column_key and extra fields in MysqlMetadata.

Removes the strange formatting of the names of constraints that aren't foreign keys. I'm unsure why that's ever been a thing, as it doesn't make much sense adding _laminas_{table name} to them?

Wraps the constraint array in a non-null check to prevent a possible null pointer exception

Converts a couple of closures to static functions for better scoping and potential performance improvements

Signed-off-by: Will Jones <will@stratagem-uk.tech>